Refactored analysis in several files
