+++ analysis-module/src/vars.py
@@ -1,0 +1,13 @@
+G_DEFAULT_SIMULATION_PATH = "/simulation_data/"
+G_DEFAULT_TOML_PATH = "/tomls/"
+
+G_DEFAULT_FIG_FILENAME = 'analysis.pdf'
+G_DEFAULT_SUMMARY_FILENAME = 'summary.json'
+G_TOPOLOGY_FILE_NAME = "network_data.json"
+
+
+G_NWAKU_IMAGE_NAME = "nim-waku"
+G_GOWAKU_IMAGE_NAME = "go-waku"
+
+G_APP_NAME = 'WLS-ANALYSIS'
+G_LOG_LEVEL = 'DEBUG'
